Disable the global and registry parts of DCell

+++ spec/support/helpers.rb
@@ -32,6 +32,7 @@ module TestNode
   end
 
   def self.stop
+    return unless @pid
     Process.kill 9, @pid
   rescue Errno::ESRCH
   ensure
